+++ src/model/tracks.rs
@@ -42,7 +42,7 @@ pub struct Track {
     pub disc_number: u8,
 
     /// The track length in milliseconds.
-    pub duration_ms: u32,
+    pub duration_ms: u64,
 
     /// Whether or not the track has explicit lyrics ( true = yes it does; false = no it does not OR unknown).
     pub explicit: bool,
@@ -115,7 +115,7 @@ pub struct SimplifiedTrack {
     pub disc_number: u8,
 
     /// The track length in milliseconds.
-    pub duration_ms: u32,
+    pub duration_ms: u64,
 
     /// Whether or not the track has explicit lyrics ( true = yes it does; false = no it does not OR unknown).
     pub explicit: bool,
